requirements.txt
langchain==0.0.308
openai
PyPDF2==3.0.1
streamlit==1.23.1
XlsxWriter==3.1.2
pydantic
openpyxl
tiktoken
faiss-cpu
fastapi
llama-cpp-python==0.1.48